requirements.txt
progressbar2==4.2.0
numpy~=1.23.0
tensorflow==2.12.0